requirements.txt
fastapi
uvicorn
python-multipart
pillow
opencv-python
transformers
torch
torchvision
sentence-transformers
faiss-cpu
pandas
numpy
requests
python-dotenv
jinja2
aiofiles
pydantic
openai
langdetect